+++ ProfilerCore/ProfilerServer.cpp
@@ -41,7 +41,7 @@ void Server::Send(DataResponse::Type type, OutputDataStream& stream)
 
 	std::string data = stream.GetData();
 
-	DataResponse response(type, data.size());
+	DataResponse response(type, (uint32)data.size());
 	socket->Send((char*)&response, sizeof(response));
 	socket->Send(data.c_str(), data.size());
 }
